requirements.txt
ipdb
scipy
transformers==4.24.0
sentencepiece==0.1.97
datasets==2.8.0
tqdm
protobuf==3.20.0
promptsource==0.2.3
evaluate
scikit-learn==0.24
wandb
gpustat
black
open-clip-torch==2.0.2